extract out ram and opcodes
